refactor(security): separate route discovery and evaluation
